requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
pandas
sqlalchemy
psycopg2-binary
apache-airflow-providers-slack
plotly
streamlit
requests